refactor(wayland): extract spawn policy from surface lifecycle
Separate spawn placement and pan behavior from surface lifecycle handling
into a dedicated module.

- move star-based spawn placement, nearby-cluster search, and spawn-pan
  queue/tick logic into `wayland/spawn.rs`
- keep `surface_lifecycle.rs` focused on surface identity, commit
  handling, node creation, and lifecycle/drop glue
- wire the new spawn module through `wayland/mod.rs` with minimal shared
  helper visibility across modules
- relocate spawn-policy tests alongside the new module to preserve
  coverage of placement and pan behavior
